+++ server/db/userQueries.js
@@ -8,9 +8,10 @@ export const userInfoQuery = (userId) => {
 };
 
 export const userProfileQuery = (userId) => {
-  return `SELECT profile_img FROM User where User.user_id = ${userId};`;
+  return `SELECT profile_img,user_name FROM User where User.user_id = ${userId};`;
 };
 
 export const myInfoQuery = (userId) => {
   return `SELECT user_name,email,profile_img FROM User where user_id = ${userId};`;
 };
+
